[MOVE] location class 이동 #291
