+++ src/main/java/FIS/iLUVit/repository/CenterRepositoryCustom.java
@@ -2,7 +2,7 @@ package FIS.iLUVit.repository;
 
 import FIS.iLUVit.dto.center.CenterDto;
 import FIS.iLUVit.dto.center.CenterRecommendDto;
-import FIS.iLUVit.domain.Location;
+import FIS.iLUVit.domain.embeddable.Location;
 import FIS.iLUVit.domain.embeddable.Area;
 import FIS.iLUVit.domain.embeddable.Theme;
 import FIS.iLUVit.domain.enumtype.KindOf;
